refactor(dashboard): move env, cluster config to apollo open api configuration properties
